#10 add testcase, move er entity location
